+++ R/ffDTW.R
@@ -5,7 +5,7 @@
 #' @usage ffDTW(X, wl = 512, length.out = 20, wn = "hanning", ovlp = 70, 
 #' bp = c(0, 22), threshold = 5, img = TRUE, parallel = 1, path = NULL, 
 #' img.suffix = "ffDTW", pb = TRUE, clip.edges = FALSE, window.type = "none", 
-#' open.end = FALSE, ...)
+#' open.end = FALSE, scale = FALSE, ...)
 #' @param  X Data frame with results containing columns for sound file name (sound.files), 
 #' selection number (selec), and start and end time of signal (start and end).
 #' The ouptut of \code{\link{manualoc}} or \code{\link{autodetec}} can be used as the input data frame. 
@@ -39,6 +39,8 @@
 #' open-ended alignments (see \code{\link[dtw]{dtw}}).
 #' @param ... Additional arguments to be passed to \code{\link{trackfreqs}} for customizing
 #' graphical output.
+#' @param scale Logical. If \code{TRUE} dominant frequency values are z-transformed using the \code{\link[base]{scale}} function, which "ignores" differences in absolute frequencies between the signals in order to focus the 
+#' comparison in the frequency contour, regardless of the pitch of signals. Default is \code{TRUE}.
 #' @return A matrix with the pairwise dissimilarity values. If img is 
 #' \code{FALSE} it also produces image files with the spectrograms of the signals listed in the 
 #' input data frame showing the location of the fundamental frequencies.
@@ -77,7 +79,7 @@
 ffDTW <- function(X, wl = 512, length.out = 20, wn = "hanning", ovlp = 70, 
                   bp = c(0, 22), threshold = 5, img = TRUE, parallel = 1, path = NULL, 
                   img.suffix = "ffDTW", pb = TRUE, clip.edges = FALSE,
-                  window.type = "none", open.end = FALSE, ...){     
+                  window.type = "none", open.end = FALSE, scale = FALSE, ...){     
     
   #stop if only 1 selection
   if(nrow(X) == 1) stop("you need more than one selection for ffDTW")
@@ -89,7 +91,10 @@ ffDTW <- function(X, wl = 512, length.out = 20, wn = "hanning", ovlp = 70,
   
   #matrix of fund freq time series
   mat <- res[,3:ncol(res)]
-    
+ 
+  if(scale)
+    mat <- t(apply(mat, 1, scale))  
+  
   #stop if NAs in matrix
     if(any(is.na(mat))) stop("missing values in frequency time series (fundamental frequency was not detected at one or both extremes of the signal)")